# more faster index on notifications table