--- a/db/migrate/20171226094803_more_faster_index_on_notifications.rb
+++ b/db/migrate/20171226094803_more_faster_index_on_notifications.rb
@@ -1,0 +1,7 @@
+class MoreFasterIndexOnNotifications < ActiveRecord::Migration[5.1]
+  def change
+    commit_db_transaction
+    add_index :notifications, [:account_id, :id], order: { id: :desc }, algorithm: :concurrently
+    remove_index :notifications, name: :index_notifications_on_id_and_account_id_and_activity_type
+  end
+end