This PR makes more faster index on notifications table.
related https://github.com/tootsuite/mastodon/pull/4750

x2.95 ～ more faster than now.

These explains were created from snapshot of friends.nico.

## very active user
x2.95 faster
### before
```
mastodon-production=> explain analyze SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = x AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT 20;
                                                                                         QUERY PLAN                                                         
                                
------------------------------------------------------------------------------------------------------------------------------------------------------------
--------------------------------
 Limit  (cost=0.56..147.46 rows=20 width=33) (actual time=0.073..0.219 rows=20 loops=1)
   ->  Index Scan using index_notifications_on_id_and_account_id_and_activity_type on notifications  (cost=0.56..851376.48 rows=115910 width=33) (actual time=0.071..0.211 rows=20 loops=1)
         Index Cond: (account_id = x)
         Filter: ((activity_type)::text = ANY ('{Mention,Status,Follow,Favourite}'::text[]))
 Planning time: 0.182 ms
 Execution time: 0.248 ms
(6 rows)
```
### after
```
mastodon-production=> explain analyze SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = x AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT 20;
                                                                          QUERY PLAN                                                                        
  
------------------------------------------------------------------------------------------------------------------------------------------------------------
--
 Limit  (cost=0.56..66.07 rows=20 width=33) (actual time=0.021..0.054 rows=20 loops=1)
   ->  Index Scan using index_notifications_20171226 on notifications  (cost=0.56..379643.68 rows=115910 width=33) (actual time=0.019..0.043 rows=20 loops=1)
         Index Cond: (account_id = x)
         Filter: ((activity_type)::text = ANY ('{Mention,Status,Follow,Favourite}'::text[]))
 Planning time: 0.265 ms
 Execution time: 0.084 ms
(6 rows)
```

## active user
x30.4 faster
### before
```
mastodon-production=> explain analyze SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = y AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT 20;
                                                                                        QUERY PLAN                                                          
                              
------------------------------------------------------------------------------------------------------------------------------------------------------------
------------------------------
 Limit  (cost=0.56..4347.50 rows=20 width=33) (actual time=1.549..2.877 rows=20 loops=1)
   ->  Index Scan using index_notifications_on_id_and_account_id_and_activity_type on notifications  (cost=0.56..494247.53 rows=2274 width=33) (actual time=1.546..2.869 rows=20 loops=1)
         Index Cond: (account_id = y)
         Filter: ((activity_type)::text = ANY ('{Mention,Status,Follow,Favourite}'::text[]))
 Planning time: 0.145 ms
 Execution time: 2.922 ms
(6 rows)
```

### after
```
mastodon-production=> explain analyze SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = y AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT 20;
                                                                        QUERY PLAN                                                                        
----------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.56..80.72 rows=20 width=33) (actual time=0.033..0.071 rows=20 loops=1)
   ->  Index Scan using index_notifications_20171226 on notifications  (cost=0.56..9114.20 rows=2274 width=33) (actual time=0.031..0.060 rows=20 loops=1)
         Index Cond: (account_id = y)
         Filter: ((activity_type)::text = ANY ('{Mention,Status,Follow,Favourite}'::text[]))
 Planning time: 0.150 ms
 Execution time: 0.096 ms
(6 rows)
```

## inactive user (worst case)
x20110.8 faster
### before
```
mastodon-production=> explain analyze SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = z AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT 20;
                                                                                          QUERY PLAN                                                        
                                  
------------------------------------------------------------------------------------------------------------------------------------------------------------
----------------------------------
 Limit  (cost=0.56..4347.50 rows=20 width=33) (actual time=56.979..1990.931 rows=20 loops=1)
   ->  Index Scan using index_notifications_on_id_and_account_id_and_activity_type on notifications  (cost=0.56..494247.53 rows=2274 width=33) (actual time=56.975..1990.906 rows=20 loops=1)
         Index Cond: (account_id = z)
         Filter: ((activity_type)::text = ANY ('{Mention,Status,Follow,Favourite}'::text[]))
 Planning time: 0.145 ms
 Execution time: 1990.972 ms
(6 rows)
```
### after
```
mastodon-production=> explain analyze SELECT  "notifications"."id", "notifications"."updated_at", "notifications"."activity_type", "notifications"."activity_id" FROM "notifications" WHERE "notifications"."account_id" = z AND "notifications"."activity_type" IN ('Mention', 'Status', 'Follow', 'Favourite') ORDER BY "notifications"."id" DESC LIMIT 20;
                                                                        QUERY PLAN                                                                        
----------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.56..80.72 rows=20 width=33) (actual time=0.030..0.073 rows=20 loops=1)
   ->  Index Scan using index_notifications_20171226 on notifications  (cost=0.56..9114.20 rows=2274 width=33) (actual time=0.028..0.062 rows=20 loops=1)
         Index Cond: (account_id = z)
         Filter: ((activity_type)::text = ANY ('{Mention,Status,Follow,Favourite}'::text[]))
 Planning time: 0.153 ms
 Execution time: 0.099 ms
(6 rows)
```