@ykzts 2017-12-26
> x20110.8 faster

Wow 😎